requirements.txt
telethon>=1.33.1
requests>=2.31.0
yt-dlp>=2025.1.1
wikipedia>=1.4.0
pytgcalls>=3.0.0
google-api-python-client>=2.153.0
pillow>=10.2.0
opencv-python>=4.9.0
pytesseract>=0.3.10
python-dotenv>=1.0.1
spotipy>=2.23.0
python-telegram-bot>=20.8
cryptography>=42.0.5
qrcode>=7.4.2
beautifulsoup4>=4.12.3
lxml>=5.2.1
html5lib>=1.1
aiohttp>=3.9.5
